Move Runes enum to api
